Use a reliable method to check for the text function
The function index exists() accesses doesn't seem to be updated within
a currently running function. This approach only worked because I tested
it with c files where the FileType autocommand was triggered twice,
resulting in the second exists() being successful.

Trying to run the function and catching the resulting exception is
a much more reliable way, since vim tries to load the autoload function
if it can't find it.